Add demonstration of the Command Pattern.
This set of classes models a remote control with a number of "programmable" on/off buttons. The remote caters for seven slots of on/off button pairs. Each pair of buttons can have a "command" assigned that knows how to perform an action on a device such as a light, poolpump, stereo, television, room fan etc.

The command pattern encapsulates actions and isolates the implementation from the client via an intermediate "invoker" object. The buttons know nothing about the device their are controlling.