Add a Primefaces UI
Add Primefaces support with a Controller and a UI that delegates to our RemoteControl class.
Add jboss-el to the project so that we can pass arguments in EL expression methods.
Add this portion to the web.xml in order for jboss-el to function:

<context-param>     
    <param-name>com.sun.faces.expressionFactory</param-name>
    <param-value>org.jboss.el.ExpressionFactoryImpl</param-value>   
</context-param>

We should not have to do this when upgrading to JSF 2.2.