Add CDI Event Bridge.
Add classes to implement the Decorator pattern in conjunction with the Observer pattern so that events in the Remote framework can be bridged to CDI events.

Add skeleton class for a bridge to JMS. This could also be implemented as a CDI event handler but we need to demonstrate the extensibility of the design.